+++ .github/workflows/tests.yaml
@@ -1,31 +1,31 @@
 name: Run Docker Compose and Test Endpoints
 
 on:
-    workflow_dispatch:
+  workflow_dispatch:
 
 jobs:
   test:
     runs-on: ubuntu-latest
 
     steps:
-    - name: Check out code
-      uses: actions/checkout@v4
+      - name: Check out code
+        uses: actions/checkout@v4
 
-    - name: Set up JDK 21
-      uses: actions/setup-java@v4
-      with:
-        distribution: 'zulu'
-        java-version: '21'
-        cache: maven
+      - name: Set up JDK 21
+        uses: actions/setup-java@v4
+        with:
+          distribution: 'zulu'
+          java-version: '21'
+          cache: maven
 
-    - name: Build with Maven
-      run: |
-        cd engine
-        mvn clean package
+      - name: Build with Maven
+        run: |
+          cd engine
+          mvn clean package
 
-    - name: Start Docker Compose
-      run: |
-        docker-compose -f docker-compose-go.yaml up -d
+      - name: Start Docker Compose
+        run: |
+          docker-compose -f docker-compose-go.yaml up -d
 
     - name: Start producing to database
       run: |
