+++ .idea/compiler.xml
@@ -6,13 +6,13 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="measurement-generator" />
         <module name="engine" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
       <module name="consumer" target="17" />
       <module name="GUI" target="17" />
+      <module name="measurement-generator" target="17" />
       <module name="product-producer" target="17" />
       <module name="Stream-Query-Engine" target="17" />
     </bytecodeTargetLevel>
