+++ docker-compose-go.yaml
@@ -37,7 +37,7 @@ services:
       - .env
     networks:
       - hdised
-    command: ["postgres", "-c", "config_file=/etc/postgresql/postgresql.conf"]
+    command: [ "postgres", "-c", "config_file=/etc/postgresql/postgresql.conf" ]
   zookeeper:
     container_name: zookeeper
     image: 'bitnami/zookeeper:latest'
@@ -60,13 +60,13 @@ services:
     networks:
       - hdised
 
-  pmm-monitoring:
-    container_name: pmm-monitoring
-    image: 'percona/pmm-server:2.41.2'
-    ports:
-      - '443:443'
-    networks:
-      - hdised
+   pmm-monitoring:
+     container_name: pmm-monitoring
+     image: 'percona/pmm-server:2.41.2'
+     ports:
+       - '443:443'
+     networks:
+       - hdised
 
 volumes:
   hdised:
