+++ engine/src/main/java/pl/polsl/hdised/engine/EngineApplication.java
@@ -6,7 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class EngineApplication {
 
-    public static void main(String... args) {
-        SpringApplication.run(EngineApplication.class, args);
-    }
+  public static void main(String... args) {
+    SpringApplication.run(EngineApplication.class, args);
+  }
 }
