+++ engine/src/main/java/pl/polsl/hdised/engine/config/KafkaTopicConfig.java
@@ -8,13 +8,13 @@ import org.springframework.kafka.config.TopicBuilder;
 @Configuration
 public class KafkaTopicConfig {
 
-    @Bean
-    public NewTopic topic() {
-        return TopicBuilder.name("topic").build();
-    }
+  @Bean
+  public NewTopic topic() {
+    return TopicBuilder.name("topic").build();
+  }
 
-    @Bean
-    public NewTopic streamTopic() {
-        return TopicBuilder.name("streamTopic").build();
-    }
+  @Bean
+  public NewTopic streamTopic() {
+    return TopicBuilder.name("streamTopic").build();
+  }
 }
