+++ engine/src/main/java/pl/polsl/hdised/engine/date/DateEntity.java
@@ -6,17 +6,16 @@ import java.util.Date;
 @Entity(name = "Date")
 public class DateEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    @Column(name = "scan_date", nullable = false, columnDefinition = "TIMESTAMP")
-    private Date scanDate;
+  @Column(name = "scan_date", nullable = false, columnDefinition = "TIMESTAMP")
+  private Date scanDate;
 
-    public DateEntity() {
-    }
+  public DateEntity() {}
 
-    public DateEntity(Date scanDate) {
-        this.scanDate = scanDate;
-    }
+  public DateEntity(Date scanDate) {
+    this.scanDate = scanDate;
+  }
 }
