+++ engine/src/main/java/pl/polsl/hdised/engine/date/DateRepository.java
@@ -2,5 +2,4 @@ package pl.polsl.hdised.engine.date;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DateRepository extends JpaRepository<DateEntity, Long> {
-}
+public interface DateRepository extends JpaRepository<DateEntity, Long> {}
