--- engine/src/main/java/pl/polsl/hdised/engine/device/DeviceDto.java
@@ -1,17 +1,0 @@
-package pl.polsl.hdised.engine.device;
-
-import java.io.Serializable;
-
-
-public class DeviceDto implements Serializable{
-
-    private final String deviceId;
-
-    public DeviceDto(String deviceId) {
-        this.deviceId = deviceId;
-    }
-
-    public String getDeviceId() {
-        return deviceId;
-    }
-}
