+++ engine/src/main/java/pl/polsl/hdised/engine/exception/EmptyMeasurementsException.java
@@ -2,7 +2,7 @@ package pl.polsl.hdised.engine.exception;
 
 public class EmptyMeasurementsException extends Exception {
 
-    public EmptyMeasurementsException() {
-        super("Measurements from device and location with provided parameters has not been found yet");
-    }
+  public EmptyMeasurementsException() {
+    super("Measurements from device and location with provided parameters has not been found yet");
+  }
 }
