+++ engine/src/main/java/pl/polsl/hdised/engine/exception/ParametersNotFoundException.java
@@ -2,7 +2,7 @@ package pl.polsl.hdised.engine.exception;
 
 public class ParametersNotFoundException extends Exception {
 
-    public ParametersNotFoundException() {
-        super("Provided parameters are invalid");
-    }
+  public ParametersNotFoundException() {
+    super("Provided parameters are invalid");
+  }
 }
