+++ engine/src/main/java/pl/polsl/hdised/engine/listener/KafkaListeners.java
@@ -2,28 +2,32 @@ package pl.polsl.hdised.engine.listener;
 
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
-import pl.polsl.hdised.engine.measurement.MeasurementDto;
-import pl.polsl.hdised.engine.streamquery.StreamQueryService;
-
+import pl.polsl.hdised.engine.measurement.model.MeasurementDto;
+import pl.polsl.hdised.engine.query.stream.StreamQueryService;
 
 @Component
 public class KafkaListeners {
 
-    private final StreamQueryService streamQueryService;
+  private final StreamQueryService streamQueryService;
 
-    public KafkaListeners(StreamQueryService streamQueryService) {
-        this.streamQueryService = streamQueryService;
-    }
+  public KafkaListeners(StreamQueryService streamQueryService) {
+    this.streamQueryService = streamQueryService;
+  }
 
-    // IF DATA WILL BE OUR CUSTOM CLASS WE WILL PASS THERE THAT CLASS
-    @KafkaListener(topics = "topic", groupId = "hdised", containerFactory = "kafkaListenerContainerFactory")
-    void listener(MeasurementDto measurementDto) {
-        this.streamQueryService.addMeasurement(measurementDto);
-    }
+  // IF DATA WILL BE OUR CUSTOM CLASS WE WILL PASS THERE THAT CLASS
+  @KafkaListener(
+      topics = "topic",
+      groupId = "hdised",
+      containerFactory = "kafkaListenerContainerFactory")
+  void listener(MeasurementDto measurementDto) {
+    this.streamQueryService.addMeasurement(measurementDto);
+  }
 
-    @KafkaListener(topics = "streamTopic", groupId = "hdised", containerFactory = "kafkaListenerContainerFactory")
-    void streamListener(MeasurementDto measurementDto) {
-        this.streamQueryService.UpdateQuery(measurementDto);
-    }
-
+  @KafkaListener(
+      topics = "streamTopic",
+      groupId = "hdised",
+      containerFactory = "kafkaListenerContainerFactory")
+  void streamListener(MeasurementDto measurementDto) {
+    this.streamQueryService.UpdateQuery(measurementDto);
+  }
 }
