--- engine/src/main/java/pl/polsl/hdised/engine/location/LocationDto.java
@@ -1,14 +1,0 @@
-package pl.polsl.hdised.engine.location;
-
-public class LocationDto {
-
-    private final String location;
-
-    public LocationDto(String location) {
-        this.location = location;
-    }
-
-    public String getLocation() {
-        return location;
-    }
-}
