+++ engine/src/main/java/pl/polsl/hdised/engine/location/model/LocationDto.java
@@ -1,0 +1,14 @@
+package pl.polsl.hdised.engine.location.model;
+
+public class LocationDto {
+
+  private final String location;
+
+  public LocationDto(String location) {
+    this.location = location;
+  }
+
+  public String getLocation() {
+    return location;
+  }
+}
