+++ engine/src/main/resources/application.properties
@@ -1,12 +1,9 @@
 server.port=8080
 spring.application.name=Engine
-
 spring.kafka.bootstrap-servers=${KAFKA_SERVER}
-
 spring.datasource.url=${DB_URL}
 spring.datasource.username=${DB_USER}
 spring.datasource.password=${DB_PASSWORD}
-
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.properties.hibernate.format_sql=true
