+++ golang-measurement-generator/go.mod
@@ -1,15 +1,15 @@
 module golang-measurement-generator
 
-go 1.21.5
+        go 1.21.5
 
-require (
-	github.com/gorilla/mux v1.8.1
-	github.com/segmentio/kafka-go v0.4.47
-)
+        require (
+        github.com/gorilla/mux v1.8.1
+        github.com/segmentio/kafka-go v0.4.47
+        )
 
-require (
-	github.com/klauspost/compress v1.17.4 // indirect
-	github.com/pierrec/lz4/v4 v4.1.15 // indirect
-	github.com/stretchr/testify v1.9.0 // indirect
-	golang.org/x/net v0.22.0 // indirect
-)
+        require (
+        github.com/klauspost/compress v1.17.4 // indirect
+        github.com/pierrec/lz4/v4 v4.1.15 // indirect
+        github.com/stretchr/testify v1.9.0 // indirect
+        golang.org/x/net v0.22.0 // indirect
+        )
